fixup! many: move VolumesAuthOptions under gadget/device
This avoids import snapd/secboot from the client package
