+++ gadget/device/encrypt_test.go
@@ -23,6 +23,7 @@ import (
 	"os"
 	"path/filepath"
 	"testing"
+	"time"
 
 	. "gopkg.in/check.v1"
 
@@ -149,4 +150,45 @@ func (s *deviceSuite) TestSealedKeysMethodWithWrongContentHappy(c *C) {
 	mth, err := device.SealedKeysMethod(root)
 	c.Check(err, IsNil)
 	c.Check(string(mth), Equals, "invalid-sealing-method")
+}
+
+func (s *deviceSuite) TestVolumesAuthOptionsValidateHappy(c *C) {
+	var opts *device.VolumesAuthOptions
+
+	// VolumesAuthOptions can be nil
+	c.Assert(opts.Validate(), IsNil)
+	// Valid kdf types
+	for _, kdfType := range []string{"argon2i", "argon2id", "pbkdf2"} {
+		opts = &device.VolumesAuthOptions{
+			Mode:       device.AuthModePassphrase,
+			Passphrase: "1234",
+			KDFType:    kdfType,
+			KDFTime:    2 * time.Second,
+		}
+		c.Assert(opts.Validate(), IsNil)
+	}
+	// KDF type and time are optional
+	opts = &device.VolumesAuthOptions{Mode: device.AuthModePassphrase, Passphrase: "1234"}
+	c.Assert(opts.Validate(), IsNil)
+}
+
+func (s *deviceSuite) TestVolumesAuthOptionsValidateError(c *C) {
+	// Bad auth mode
+	opts := &device.VolumesAuthOptions{Mode: "bad-mode", Passphrase: "1234"}
+	c.Assert(opts.Validate(), ErrorMatches, `invalid authentication mode "bad-mode", only "passphrase" and "pin" modes are supported`)
+	// Empty passphrase
+	opts = &device.VolumesAuthOptions{Mode: device.AuthModePassphrase}
+	c.Assert(opts.Validate(), ErrorMatches, "passphrase cannot be empty")
+	// PIN mode not implemented yet
+	opts = &device.VolumesAuthOptions{Mode: device.AuthModePIN}
+	c.Assert(opts.Validate(), ErrorMatches, `"pin" authentication mode is not implemented`)
+	// PIN mode + custom kdf type
+	opts = &device.VolumesAuthOptions{Mode: device.AuthModePIN, KDFType: "argon2i"}
+	c.Assert(opts.Validate(), ErrorMatches, `"pin" authentication mode does not support custom kdf types`)
+	// Bad kdf type
+	opts = &device.VolumesAuthOptions{Mode: device.AuthModePassphrase, Passphrase: "1234", KDFType: "bad-type"}
+	c.Assert(opts.Validate(), ErrorMatches, `invalid kdf type "bad-type", only "argon2i", "argon2id" and "pbkdf2" are supported`)
+	// Negative kdf time
+	opts = &device.VolumesAuthOptions{Mode: device.AuthModePassphrase, Passphrase: "1234", KDFTime: -1}
+	c.Assert(opts.Validate(), ErrorMatches, "kdf time cannot be negative")
 }
